+++ src/main/java/org/geowe/service/constraints/SimpleGeometry.java
@@ -24,6 +24,8 @@ import java.lang.annotation.Target;
 import javax.validation.Constraint;
 import javax.validation.Payload;
 
+import org.geowe.service.constraints.validator.SimpleGeometryValidator;
+
 @Target({ ElementType.FIELD })
 @Retention(RetentionPolicy.RUNTIME)
 @Constraint(validatedBy = SimpleGeometryValidator.class)