+++ src/main/java/org/geowe/service/constraints/TopologyGroup.java
@@ -1,5 +1,0 @@
-package org.geowe.service.constraints;
-
-public interface TopologyGroup {
-
-}