+++ src/main/java/org/geowe/service/constraints/group/TopologyGroup.java
@@ -1,0 +1,5 @@
+package org.geowe.service.constraints.group;
+
+public interface TopologyGroup {
+
+}