+++ src/main/java/org/geowe/service/model/FlatGeometry.java
@@ -21,9 +21,11 @@ import javax.validation.constraints.NotNull;
 
 import org.geowe.service.constraints.DuplicateVertex;
 import org.geowe.service.constraints.SimpleGeometry;
-import org.geowe.service.constraints.SimplicityGroup;
-import org.geowe.service.constraints.TopologyGroup;
 import org.geowe.service.constraints.ValidTopology;
+import org.geowe.service.constraints.VertexOrientation;
+import org.geowe.service.constraints.group.SimplicityGroup;
+import org.geowe.service.constraints.group.TopologyGroup;
+import org.geowe.service.constraints.group.VertexOrientationGroup;
 import org.hibernate.validator.constraints.NotBlank;
 
 /**
@@ -46,6 +48,7 @@ public class FlatGeometry implements Serializable{
 	@ValidTopology(groups={TopologyGroup.class})
 	@SimpleGeometry(groups={SimplicityGroup.class})
 	@DuplicateVertex(groups={TopologyGroup.class})
+	@VertexOrientation(groups={VertexOrientationGroup.class})
 	private String wkt;
 
 	public FlatGeometry(){