+++ CHANGELOG.md
@@ -12,8 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.9.4]SNAPSHOT - 2017-04-05
 
 ### Added
-- Queryparam (validate) for Geometry validation resource, to choose validation type to perform
-- Suported validations: JTS valid, JTS simple, repeated coordinates
+- QueryParam (validate) for Geometry validation resource, to choose validation type to perform
+- Supported validations: JTS valid, JTS simple, repeated coordinates, vertices orientation (SFS compliance).
 
 ## [0.9.0]SNAPSHOT - 2017-03-31